requirements.txt
torch
torchvision
numpy
opencv-python
matplotlib
pillow
tqdm
requests